+++ tests/errors/semantic/blocking/UNDEFINED_LAMBDA_FUNCTION.py
@@ -1,3 +1,0 @@
-# pylint: disable=missing-function-docstring, missing-module-docstring
-a = lambda x: 1+b()
-